+++ web/src/main/java/org/cmc/curtaincall/web/party/PartyController.java
@@ -3,24 +3,14 @@ package org.cmc.curtaincall.web.party;
 import jakarta.validation.constraints.Size;
 import lombok.RequiredArgsConstructor;
 import org.cmc.curtaincall.domain.member.MemberId;
-import org.cmc.curtaincall.domain.party.PartyCategory;
-import org.cmc.curtaincall.domain.party.PartyId;
-import org.cmc.curtaincall.domain.party.dao.PartyDao;
-import org.cmc.curtaincall.domain.party.request.PartySearchParam;
-import org.cmc.curtaincall.domain.party.response.PartyDetailResponse;
-import org.cmc.curtaincall.domain.party.response.PartyResponse;
 import org.cmc.curtaincall.web.common.response.IdResult;
-import org.cmc.curtaincall.web.common.response.ListResult;
 import org.cmc.curtaincall.web.exception.EntityAccessDeniedException;
 import org.cmc.curtaincall.web.party.request.PartyCreate;
 import org.cmc.curtaincall.web.party.request.PartyEdit;
 import org.cmc.curtaincall.web.party.response.PartyParticipatedResponse;
 import org.cmc.curtaincall.web.security.LoginMemberId;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.SliceImpl;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.SortDefault;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,26 +21,6 @@ import java.util.List;
 public class PartyController {
 
     private final PartyService partyService;
-
-    private final PartyDao partyDao;
-
-    @GetMapping("/parties")
-    public ListResult<PartyResponse> getPartyList(
-            @SortDefault(sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable,
-            @RequestParam PartyCategory category) {
-        return new ListResult<>(partyDao.getList(pageable, category));
-    }
-
-    @GetMapping("/search/party")
-    public ListResult<PartyResponse> searchParty(
-            Pageable pageable, @ModelAttribute @Validated PartySearchParam partySearchParam) {
-        return new ListResult<>(partyDao.search(pageable, partySearchParam));
-    }
-
-    @GetMapping("/parties/{partyId}")
-    public PartyDetailResponse getPartyDetail(@PathVariable Long partyId) {
-        return partyDao.getDetail(new PartyId(partyId));
-    }
 
     @PostMapping("/parties")
     public IdResult<Long> createParty(@RequestBody @Validated PartyCreate partyCreate) {
