+++ web/src/test/java/org/cmc/curtaincall/web/controller/MemberControllerDocsTest.java
@@ -2,10 +2,6 @@ package org.cmc.curtaincall.web.controller;
 
 import org.cmc.curtaincall.domain.lostitem.LostItemType;
 import org.cmc.curtaincall.domain.member.MemberDeleteReason;
-import org.cmc.curtaincall.domain.party.PartyCategory;
-import org.cmc.curtaincall.domain.party.dao.PartyDao;
-import org.cmc.curtaincall.domain.party.response.PartyParticipationResponse;
-import org.cmc.curtaincall.domain.party.response.PartyRecruitmentResponse;
 import org.cmc.curtaincall.web.common.AbstractWebTest;
 import org.cmc.curtaincall.web.common.RestDocsAttribute;
 import org.cmc.curtaincall.web.common.response.BooleanResult;
@@ -63,9 +59,6 @@ class MemberControllerDocsTest extends AbstractWebTest {
 
     @MockBean
     private LostItemService lostItemService;
-
-    @MockBean
-    private PartyDao partyDao;
 
     @Test
     void getNicknameDuplicate_Docs() throws Exception {
@@ -193,140 +186,6 @@ class MemberControllerDocsTest extends AbstractWebTest {
                                 fieldWithPath("nickname").description("닉네임")
                                         .attributes(RestDocsAttribute.constraint("min = 2, max = 15")),
                                 fieldWithPath("imageId").description("이미지 ID").optional()
-                        )
-                ));
-    }
-
-    @Test
-    void getRecruitmentList_Docs() throws Exception {
-        // given
-        var partyResponse = PartyRecruitmentResponse.builder()
-                .id(10L)
-                .title("공연 같이 보실분~")
-                .content("저랑 같이 봐요~")
-                .curMemberNum(2)
-                .maxMemberNum(5)
-                .showAt(LocalDateTime.of(2023, 4, 28, 19, 30))
-                .createdAt(LocalDateTime.of(2023, 4, 28, 11, 12, 28))
-                .category(PartyCategory.WATCHING)
-                .showId("PF220846")
-                .showName("잘자요, 엄마 [청주]")
-                .showPoster("post-image-url")
-                .facilityId("FC000182")
-                .facilityName("예술나눔 터 (예술나눔 터)")
-                .build();
-        given(partyDao.getRecruitmentList(any(), any(), any()))
-                .willReturn(List.of(partyResponse));
-
-        // expected
-        mockMvc.perform(get("/members/{memberId}/recruitments", 2L)
-                        .header(HttpHeaders.AUTHORIZATION, "Bearer ACCESS_TOKEN")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .param("page", "0")
-                        .param("size", "20")
-                        .param("category", PartyCategory.WATCHING.name())
-                )
-                .andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("member-get-recruitment-list",
-                        requestHeaders(
-                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 필요")
-                        ),
-                        pathParameters(
-                                parameterWithName("memberId").description("회원 ID")
-                        ),
-                        queryParameters(
-                                parameterWithName("page").description("페이지"),
-                                parameterWithName("size").description("페이지 사이즈").optional(),
-                                parameterWithName("category").description("카테고리")
-                                        .attributes(RestDocsAttribute.type(PartyCategory.class))
-                                        .optional()
-                        ),
-                        responseFields(
-                                beneathPath("content[]").withSubsectionId("content"),
-                                fieldWithPath("id").description("파티 ID"),
-                                fieldWithPath("title").description("제목"),
-                                fieldWithPath("content").description("내용"),
-                                fieldWithPath("curMemberNum").description("현재 참여 인원 수"),
-                                fieldWithPath("maxMemberNum").description("최대 참여 인원 수"),
-                                fieldWithPath("showAt").description("공연 일시"),
-                                fieldWithPath("createdAt").description("작성 일시"),
-                                fieldWithPath("category").type(PartyCategory.class.getSimpleName())
-                                        .description("카테고리"),
-                                fieldWithPath("showId").description("공연 ID"),
-                                fieldWithPath("showName").description("공연 이름"),
-                                fieldWithPath("showPoster").description("공연 포스터"),
-                                fieldWithPath("facilityId").description("공연 시설 ID"),
-                                fieldWithPath("facilityName").description("공연 시설 이름")
-                        )
-                ));
-    }
-
-    @Test
-    void getParticipationList_Docs() throws Exception {
-        // given
-        var partyResponse = PartyParticipationResponse.builder()
-                .id(10L)
-                .title("공연 같이 보실분~")
-                .content("저랑 같이 봐요~")
-                .curMemberNum(2)
-                .maxMemberNum(5)
-                .showAt(LocalDateTime.of(2023, 4, 28, 19, 30))
-                .createdAt(LocalDateTime.of(2023, 4, 28, 11, 12, 28))
-                .category(PartyCategory.WATCHING)
-                .creatorId(2L)
-                .creatorNickname("고라파덕")
-                .creatorImageUrl("creator-image-url")
-                .showId("PF220846")
-                .showName("잘자요, 엄마 [청주]")
-                .showPoster("post-image-url")
-                .facilityId("FC000182")
-                .facilityName("예술나눔 터 (예술나눔 터)")
-                .build();
-        given(partyDao.getParticipationList(any(), any(), any()))
-                .willReturn(List.of(partyResponse));
-
-        // expected
-        mockMvc.perform(get("/members/{memberId}/participations", 2L)
-                        .header(HttpHeaders.AUTHORIZATION, "Bearer ACCESS_TOKEN")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .param("page", "0")
-                        .param("size", "20")
-                        .param("category", PartyCategory.WATCHING.name())
-                )
-                .andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("member-get-participation-list",
-                        requestHeaders(
-                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 필요")
-                        ),
-                        pathParameters(
-                                parameterWithName("memberId").description("회원 ID")
-                        ),
-                        queryParameters(
-                                parameterWithName("page").description("페이지"),
-                                parameterWithName("size").description("페이지 사이즈").optional(),
-                                parameterWithName("category").description("카테고리").optional()
-                        ),
-                        responseFields(
-                                beneathPath("content[]").withSubsectionId("content"),
-                                fieldWithPath("id").description("파티 ID"),
-                                fieldWithPath("title").description("제목"),
-                                fieldWithPath("content").description("내용"),
-                                fieldWithPath("curMemberNum").description("현재 참여 인원 수"),
-                                fieldWithPath("maxMemberNum").description("최대 참여 인원 수"),
-                                fieldWithPath("showAt").description("공연 일시"),
-                                fieldWithPath("createdAt").description("작성 일시"),
-                                fieldWithPath("category").type(PartyCategory.class.getSimpleName())
-                                        .description("카테고리"),
-                                fieldWithPath("creatorId").description("작성자 ID"),
-                                fieldWithPath("creatorNickname").description("작성자 닉네임"),
-                                fieldWithPath("creatorImageUrl").description("작성자 이미지 URL").optional(),
-                                fieldWithPath("showId").description("공연 ID"),
-                                fieldWithPath("showName").description("공연 이름"),
-                                fieldWithPath("showPoster").description("공연 포스터"),
-                                fieldWithPath("facilityId").description("공연 시설 ID"),
-                                fieldWithPath("facilityName").description("공연 시설 이름")
                         )
                 ));
     }
