+++ web/src/test/java/org/cmc/curtaincall/web/party/PartyQueryControllerDocsTest.java
@@ -1,0 +1,363 @@
+package org.cmc.curtaincall.web.party;
+
+import org.cmc.curtaincall.domain.party.PartyCategory;
+import org.cmc.curtaincall.domain.party.dao.PartyDao;
+import org.cmc.curtaincall.domain.party.response.PartyDetailResponse;
+import org.cmc.curtaincall.domain.party.response.PartyParticipationResponse;
+import org.cmc.curtaincall.domain.party.response.PartyRecruitmentResponse;
+import org.cmc.curtaincall.domain.party.response.PartyResponse;
+import org.cmc.curtaincall.web.common.AbstractWebTest;
+import org.cmc.curtaincall.web.common.RestDocsAttribute;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import static org.cmc.curtaincall.web.common.RestDocsAttribute.type;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
+import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
+import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
+import static org.springframework.restdocs.payload.PayloadDocumentation.*;
+import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
+import static org.springframework.restdocs.request.RequestDocumentation.*;
+import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@WebMvcTest(PartyQueryController.class)
+class PartyQueryControllerDocsTest extends AbstractWebTest {
+
+    @MockBean
+    private PartyDao partyDao;
+
+    @Test
+    void getPartyList_Docs() throws Exception {
+        // given
+        PartyResponse partyResponse = PartyResponse.builder()
+                .id(10L)
+                .title("공연 같이 보실분~")
+                .curMemberNum(2)
+                .maxMemberNum(5)
+                .showAt(LocalDateTime.of(2023, 4, 28, 19, 30))
+                .createdAt(LocalDateTime.of(2023, 4, 28, 11, 12, 28))
+                .category(PartyCategory.WATCHING)
+                .creatorId(2L)
+                .creatorNickname("고라파덕")
+                .creatorImageUrl("creator-image-url")
+                .showId("PF220846")
+                .showName("잘자요, 엄마 [청주]")
+                .showPoster("post-image-url")
+                .facilityId("FC000182")
+                .facilityName("예술나눔 터 (예술나눔 터)")
+                .build();
+        given(partyDao.getList(any(), any())).willReturn(List.of(partyResponse));
+
+        // expected
+        mockMvc.perform(get("/parties")
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer ACCESS_TOKEN")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .param("page", "0")
+                        .param("size", "20")
+                        .param("category", PartyCategory.WATCHING.name())
+                )
+                .andExpect(status().isOk())
+                .andDo(print())
+                .andDo(document("party-get-party-list",
+                        requestHeaders(
+                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 필요")
+                        ),
+                        queryParameters(
+                                parameterWithName("page").description("페이지"),
+                                parameterWithName("size").description("페이지 사이즈").optional(),
+                                parameterWithName("category").description("카테고리")
+                        ),
+                        responseFields(
+                                beneathPath("content[]").withSubsectionId("content"),
+                                fieldWithPath("id").description("파티 ID"),
+                                fieldWithPath("title").description("제목"),
+                                fieldWithPath("curMemberNum").description("현재 참여 인원 수"),
+                                fieldWithPath("maxMemberNum").description("최대 참여 인원 수"),
+                                fieldWithPath("showAt").description("공연 일시").optional(),
+                                fieldWithPath("createdAt").description("작성 일시"),
+                                fieldWithPath("category").type(PartyCategory.class.getSimpleName())
+                                        .description("카테고리"),
+                                fieldWithPath("creatorId").description("작성자 ID"),
+                                fieldWithPath("creatorNickname").description("작성자 닉네임"),
+                                fieldWithPath("creatorImageUrl").description("작성자 이미지 URL").optional(),
+                                fieldWithPath("showId").description("공연 ID").optional(),
+                                fieldWithPath("showName").description("공연 이름").optional(),
+                                fieldWithPath("showPoster").description("공연 포스터").optional(),
+                                fieldWithPath("facilityId").description("공연 시설 ID").optional(),
+                                fieldWithPath("facilityName").description("공연 시설 이름").optional()
+                        )
+                ));
+    }
+
+
+    @Test
+    void searchParty_Docs() throws Exception {
+        // given
+        PartyResponse partyResponse = PartyResponse.builder()
+                .id(10L)
+                .title("공연 같이 보실분~")
+                .curMemberNum(2)
+                .maxMemberNum(5)
+                .showAt(LocalDateTime.of(2023, 4, 28, 19, 30))
+                .createdAt(LocalDateTime.of(2023, 4, 28, 11, 12, 28))
+                .category(PartyCategory.WATCHING)
+                .creatorId(2L)
+                .creatorNickname("고라파덕")
+                .creatorImageUrl("creator-image-url")
+                .showId("PF220846")
+                .showName("잘자요, 엄마 [청주]")
+                .showPoster("post-image-url")
+                .facilityId("FC000182")
+                .facilityName("예술나눔 터 (예술나눔 터)")
+                .build();
+        given(partyDao.search(any(), any())).willReturn(List.of(partyResponse));
+
+        // expected
+        mockMvc.perform(get("/search/party")
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer ACCESS_TOKEN")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .param("page", "0")
+                        .param("size", "20")
+                        .param("category", PartyCategory.WATCHING.name())
+                        .param("keyword", "잘자요")
+                )
+                .andExpect(status().isOk())
+                .andDo(print())
+                .andDo(document("party-search",
+                        requestHeaders(
+                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 필요")
+                        ),
+                        queryParameters(
+                                parameterWithName("page").description("페이지"),
+                                parameterWithName("size").description("페이지 사이즈").optional(),
+                                parameterWithName("category").description("카테고리")
+                                        .attributes(type(PartyCategory.class)),
+                                parameterWithName("keyword").description("검색 키워드")
+                        ),
+                        responseFields(
+                                beneathPath("content[]").withSubsectionId("content"),
+                                fieldWithPath("id").description("파티 ID"),
+                                fieldWithPath("title").description("제목"),
+                                fieldWithPath("curMemberNum").description("현재 참여 인원 수"),
+                                fieldWithPath("maxMemberNum").description("최대 참여 인원 수"),
+                                fieldWithPath("showAt").description("공연 일시").optional(),
+                                fieldWithPath("createdAt").description("작성 일시"),
+                                fieldWithPath("category").type(PartyCategory.class.getSimpleName())
+                                        .description("카테고리"),
+                                fieldWithPath("creatorId").description("작성자 ID"),
+                                fieldWithPath("creatorNickname").description("작성자 닉네임"),
+                                fieldWithPath("creatorImageUrl").description("작성자 이미지 URL").optional(),
+                                fieldWithPath("showId").description("공연 ID").optional(),
+                                fieldWithPath("showName").description("공연 이름").optional(),
+                                fieldWithPath("showPoster").description("공연 포스터").optional(),
+                                fieldWithPath("facilityId").description("공연 시설 ID").optional(),
+                                fieldWithPath("facilityName").description("공연 시설 이름").optional()
+                        )
+                ));
+    }
+
+
+    @Test
+    void getPartyDetail_Docs() throws Exception {
+        // given
+        PartyDetailResponse partyDetailResponse = PartyDetailResponse.builder()
+                .id(10L)
+                .title("공연 같이 보실분~")
+                .content("저랑 같이 봐요~")
+                .category(PartyCategory.WATCHING)
+                .curMemberNum(2)
+                .maxMemberNum(5)
+                .showAt(LocalDateTime.of(2023, 4, 28, 19, 30))
+                .createdAt(LocalDateTime.of(2023, 4, 28, 11, 12, 28))
+                .creatorId(2L)
+                .creatorNickname("고라파덕")
+                .creatorImageUrl("creator-image-url")
+                .showId("PF220846")
+                .showName("잘자요, 엄마 [청주]")
+                .facilityId("FC000182")
+                .facilityName("예술나눔 터 (예술나눔 터)")
+                .build();
+        given(partyDao.getDetail(any())).willReturn(partyDetailResponse);
+
+        // expected
+        mockMvc.perform(get("/parties/{partyId}", 10L)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer ACCESS_TOKEN")
+                )
+                .andExpect(status().isOk())
+                .andDo(print())
+                .andDo(document("party-get-party-detail",
+                        requestHeaders(
+                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 필요")
+                        ),
+                        pathParameters(
+                                parameterWithName("partyId").description("공연 아이디")
+                        ),
+                        responseFields(
+                                fieldWithPath("id").description("파티 ID"),
+                                fieldWithPath("title").description("제목"),
+                                fieldWithPath("content").description("내용"),
+                                fieldWithPath("category").type(PartyCategory.class.getSimpleName())
+                                        .description("카테고리"),
+                                fieldWithPath("curMemberNum").description("현재 참여 인원 수"),
+                                fieldWithPath("maxMemberNum").description("최대 참여 인원 수"),
+                                fieldWithPath("showAt").description("공연 일시").optional(),
+                                fieldWithPath("createdAt").description("작성 일시"),
+                                fieldWithPath("creatorId").description("작성자 ID"),
+                                fieldWithPath("creatorNickname").description("작성자 닉네임"),
+                                fieldWithPath("creatorImageUrl").description("작성자 이미지 URL").optional(),
+                                fieldWithPath("showId").description("공연 ID").optional(),
+                                fieldWithPath("showName").description("공연 이름").optional(),
+                                fieldWithPath("showPoster").description("공연 포스터").optional(),
+                                fieldWithPath("facilityId").description("공연 시설 ID").optional(),
+                                fieldWithPath("facilityName").description("공연 시설 이름").optional()
+                        )
+                ));
+    }
+
+    @Test
+    void getRecruitmentList_Docs() throws Exception {
+        // given
+        var partyResponse = PartyRecruitmentResponse.builder()
+                .id(10L)
+                .title("공연 같이 보실분~")
+                .content("저랑 같이 봐요~")
+                .curMemberNum(2)
+                .maxMemberNum(5)
+                .showAt(LocalDateTime.of(2023, 4, 28, 19, 30))
+                .createdAt(LocalDateTime.of(2023, 4, 28, 11, 12, 28))
+                .category(PartyCategory.WATCHING)
+                .showId("PF220846")
+                .showName("잘자요, 엄마 [청주]")
+                .showPoster("post-image-url")
+                .facilityId("FC000182")
+                .facilityName("예술나눔 터 (예술나눔 터)")
+                .build();
+        given(partyDao.getRecruitmentList(any(), any(), any()))
+                .willReturn(List.of(partyResponse));
+
+        // expected
+        mockMvc.perform(get("/members/{memberId}/recruitments", 2L)
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer ACCESS_TOKEN")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .param("page", "0")
+                        .param("size", "20")
+                        .param("category", PartyCategory.WATCHING.name())
+                )
+                .andExpect(status().isOk())
+                .andDo(print())
+                .andDo(document("party-get-recruitment-list",
+                        requestHeaders(
+                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 필요")
+                        ),
+                        pathParameters(
+                                parameterWithName("memberId").description("회원 ID")
+                        ),
+                        queryParameters(
+                                parameterWithName("page").description("페이지"),
+                                parameterWithName("size").description("페이지 사이즈").optional(),
+                                parameterWithName("category").description("카테고리")
+                                        .attributes(RestDocsAttribute.type(PartyCategory.class))
+                                        .optional()
+                        ),
+                        responseFields(
+                                beneathPath("content[]").withSubsectionId("content"),
+                                fieldWithPath("id").description("파티 ID"),
+                                fieldWithPath("title").description("제목"),
+                                fieldWithPath("content").description("내용"),
+                                fieldWithPath("curMemberNum").description("현재 참여 인원 수"),
+                                fieldWithPath("maxMemberNum").description("최대 참여 인원 수"),
+                                fieldWithPath("showAt").description("공연 일시"),
+                                fieldWithPath("createdAt").description("작성 일시"),
+                                fieldWithPath("category").type(PartyCategory.class.getSimpleName())
+                                        .description("카테고리"),
+                                fieldWithPath("showId").description("공연 ID"),
+                                fieldWithPath("showName").description("공연 이름"),
+                                fieldWithPath("showPoster").description("공연 포스터"),
+                                fieldWithPath("facilityId").description("공연 시설 ID"),
+                                fieldWithPath("facilityName").description("공연 시설 이름")
+                        )
+                ));
+    }
+
+    @Test
+    void getParticipationList_Docs() throws Exception {
+        // given
+        var partyResponse = PartyParticipationResponse.builder()
+                .id(10L)
+                .title("공연 같이 보실분~")
+                .content("저랑 같이 봐요~")
+                .curMemberNum(2)
+                .maxMemberNum(5)
+                .showAt(LocalDateTime.of(2023, 4, 28, 19, 30))
+                .createdAt(LocalDateTime.of(2023, 4, 28, 11, 12, 28))
+                .category(PartyCategory.WATCHING)
+                .creatorId(2L)
+                .creatorNickname("고라파덕")
+                .creatorImageUrl("creator-image-url")
+                .showId("PF220846")
+                .showName("잘자요, 엄마 [청주]")
+                .showPoster("post-image-url")
+                .facilityId("FC000182")
+                .facilityName("예술나눔 터 (예술나눔 터)")
+                .build();
+        given(partyDao.getParticipationList(any(), any(), any()))
+                .willReturn(List.of(partyResponse));
+
+        // expected
+        mockMvc.perform(get("/members/{memberId}/participations", 2L)
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer ACCESS_TOKEN")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .param("page", "0")
+                        .param("size", "20")
+                        .param("category", PartyCategory.WATCHING.name())
+                )
+                .andExpect(status().isOk())
+                .andDo(print())
+                .andDo(document("party-get-participation-list",
+                        requestHeaders(
+                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 필요")
+                        ),
+                        pathParameters(
+                                parameterWithName("memberId").description("회원 ID")
+                        ),
+                        queryParameters(
+                                parameterWithName("page").description("페이지"),
+                                parameterWithName("size").description("페이지 사이즈").optional(),
+                                parameterWithName("category").description("카테고리").optional()
+                        ),
+                        responseFields(
+                                beneathPath("content[]").withSubsectionId("content"),
+                                fieldWithPath("id").description("파티 ID"),
+                                fieldWithPath("title").description("제목"),
+                                fieldWithPath("content").description("내용"),
+                                fieldWithPath("curMemberNum").description("현재 참여 인원 수"),
+                                fieldWithPath("maxMemberNum").description("최대 참여 인원 수"),
+                                fieldWithPath("showAt").description("공연 일시"),
+                                fieldWithPath("createdAt").description("작성 일시"),
+                                fieldWithPath("category").type(PartyCategory.class.getSimpleName())
+                                        .description("카테고리"),
+                                fieldWithPath("creatorId").description("작성자 ID"),
+                                fieldWithPath("creatorNickname").description("작성자 닉네임"),
+                                fieldWithPath("creatorImageUrl").description("작성자 이미지 URL").optional(),
+                                fieldWithPath("showId").description("공연 ID"),
+                                fieldWithPath("showName").description("공연 이름"),
+                                fieldWithPath("showPoster").description("공연 포스터"),
+                                fieldWithPath("facilityId").description("공연 시설 ID"),
+                                fieldWithPath("facilityName").description("공연 시설 이름")
+                        )
+                ));
+    }
+}
